ci: change travis build to trusty ubuntu distro
There is a known issue with travis and moodle-plugin-ci mustache, this avoids that issue.
The plugin is reliant on java openjdk-8-jre, whereas xenial and bionic ship with higher versions.
By default, java therefore refers to a higher version and mustache template linting fails.
This patch is a workaround for that behaviour until the issue with the plugin is resolved.
(See https://github.com/blackboard-open-source/moodle-plugin-ci/issues/91\)