Issue#44: Move Pricing Client creation to provisioning
There was a disconnect between the creation of the Pricing Client instance and that of other AWS clients.
The creation of the AWS Pricing Client was moved to the provisioning class in line with the other AWS Clients.
Additionally, tests for the ETS pricing client were updated to reflect the change, as an AWS Pricing Client dependency is required.