Issue#89: Updated transcode cost calculation to reflect conversion logic
Changes have been made which ensure that video only media is not passed to AWS for conversion to audio and vice versa, therefore pricing calculation has been amended to reflect this.
Audio stream and video stream count has been added to the calculation, so that pricing more accurately reflects the transcoding conducted.
Other changes had also broken the testing for the pricing calculator, seperate fixtures are now in place for the pricing calculator and tests have been amended to reflect these.