Issue#94: Only pass compatible presets with metadata
Added parsing to the conversion class to ensure that presets included in metadata are only those which are compatible with the container type of file.

This will avoid the transcoder trigger lamdba creating AWS ETS jobs which return errors due when attempting to transcode an input file with an incompatible preset.

Resolves #94